requirements.txt
azure-functions>=1.12.0
azure-functions-durable>=1.2.0
requests>=2.28.0
azure-core>=1.24.0
aiohttp>=3.8.0
azure-eventhub>=5.11.2
